+++ preview_generator/preview/builder/image__pillow.py
@@ -211,7 +211,7 @@ class PillowImageConvertStrategyFactory(object):
 
 class ImagePreviewBuilderPillow(ImagePreviewBuilder):
     """WARNING : This builder is deprecated, prefer ImagePreviewBuilderWand instead which
-        support the same list of format."""
+    support the same list of format."""
 
     weight = 20