+++ preview_generator/preview/builder_factory.py
@@ -91,7 +91,6 @@ class PreviewBuilderFactory(object):
         if force or not self.builders_loaded:
             builder_folder = get_builder_folder_name()
             builder_modules = get_builder_modules(builder_folder)
-
             for module_name in builder_modules:
                 try:
                     import_builder_module(module_name)
@@ -110,7 +109,7 @@ class PreviewBuilderFactory(object):
                     pass
                 elif cls.__name__ in ("ImagePreviewBuilderPillow", "ImagePreviewBuilderIMConvert"):
                     self.logger.info(
-                        "%s builder is deprecated and is not registered by default. "
+                        "{} builder is deprecated and is not registered by default. "
                         "Consider using ImagePreviewBuilderIMConvert instead".format(cls.__name__)
                     )
                 else:
@@ -141,6 +140,7 @@ class PreviewBuilderFactory(object):
             # like automatic travis test with Ubuntu 14.04.5 LTS,
             # where ImagePreviewBuilderIMConvert pretend to
             # be able to deal with application/octet-stream mimetype
+
             for mimetype in builder.get_supported_mimetypes():
                 if mimetype == "application/octet-stream":
                     self.logger.critical(