+++ tests/builders/test_image_wand_builder.py
@@ -1,7 +1,8 @@
 import os
+
 from PIL import Image
+
 from preview_generator.preview.builder.image__wand import ImagePreviewBuilderWand
-from preview_generator.preview.builder.image__imconvert import ImagePreviewBuilderIMConvert
 from preview_generator.utils import ImgDims
 
 CURRENT_DIR = os.path.dirname(os.path.abspath(__file__))
@@ -22,7 +23,7 @@ def test_build_jpeg_preview() -> None:
         cache_path=CACHE_DIR,
         page_id=-1,
         size=size,
-        extension=extension
+        extension=extension,
     )
     preview_name = preview_name + extension
     dest_path = os.path.join(CACHE_DIR, preview_name)