+++ tests/input/pdf/test_pdf.py
@@ -6,19 +6,18 @@ import shutil
 import typing
 
 from PIL import Image
-from PyPDF2 import PdfFileReader
-import PyPDF2.utils
 import pytest
 
 from preview_generator.exception import UnavailablePreviewType
 from preview_generator.manager import PreviewManager
+from preview_generator.preview.builder.pdf__poppler_utils import PdfPreviewBuilderPopplerUtils
 from tests import test_utils
 
 CURRENT_DIR = os.path.dirname(os.path.abspath(__file__))
 CACHE_DIR = "/tmp/preview-generator-tests/cache"
 PDF_FILE_PATH = os.path.join(CURRENT_DIR, "the_pdf.pdf")
 PDF_FILE_PATH__ENCRYPTED = os.path.join(CURRENT_DIR, "the_pdf.encrypted.pdf")
-PDF_FILE_PATH__A4 = os.path.join(CURRENT_DIR, "qpdfconvert.pdf")
+PDF_FILE_PATH__A4 = os.path.join(CURRENT_DIR, "pdfconvert.pdf")
 
 
 def setup_function(function: typing.Callable) -> None:
@@ -41,10 +40,6 @@ def test_to_jpeg() -> None:
 
 
 def test_to_jpeg__encrypted_pdf() -> None:
-    with pytest.raises(PyPDF2.utils.PdfReadError):  #  ensure file is encrpyted
-        pdf = PdfFileReader(PDF_FILE_PATH__ENCRYPTED)
-        pdf.getPage(0)
-
     manager = PreviewManager(cache_folder_path=CACHE_DIR, create_folder=True)
     assert manager.has_jpeg_preview(file_path=PDF_FILE_PATH) is True
     path_to_file = manager.get_jpeg_preview(
@@ -101,15 +96,23 @@ def test_to_pdf_one_page() -> None:
     assert os.path.exists(path_0) is True
     assert os.path.getsize(path_0) > 1000  # verify if the size of the pdf refer to a normal content
     assert re.match(test_utils.CACHE_FILE_PATH_PATTERN_WITH_PAGE__PDF, path_0)
-    pdf = PdfFileReader(open(path_0, "rb"))
-    assert pdf.getNumPages() == 1
+    assert (
+        PdfPreviewBuilderPopplerUtils().get_page_number(
+            cache_path=CACHE_DIR, preview_name="test", file_path=path_0
+        )
+        == 1
+    )
 
     path_1 = manager.get_pdf_preview(file_path=PDF_FILE_PATH, page=1, force=True)
     assert os.path.exists(path_1) is True
     assert os.path.getsize(path_1) > 1000  # verify if the size of the pdf refer to a normal content
     assert re.match(test_utils.CACHE_FILE_PATH_PATTERN_WITH_PAGE__PDF, path_1)
-    pdf = PdfFileReader(open(path_1, "rb"))
-    assert pdf.getNumPages() == 1
+    assert (
+        PdfPreviewBuilderPopplerUtils().get_page_number(
+            cache_path=CACHE_DIR, preview_name="test", file_path=path_0
+        )
+        == 1
+    )
 
 
 def test_algorithm4() -> None:
@@ -118,7 +121,7 @@ def test_algorithm4() -> None:
     path_to_file = manager.get_jpeg_preview(file_path=PDF_FILE_PATH__A4, force=True)
     with Image.open(path_to_file) as jpeg:
         assert jpeg.height == 256
-        assert jpeg.width in range(180, 182)
+        assert jpeg.width in range(180, 183)
 
 
 def test_get_nb_page() -> None: